refactor: general components placed inside _components
