+++ src/app/lembaga/profil-lembaga/[lembagaId]/page.tsx
@@ -23,7 +23,7 @@ import { CalendarIcon } from 'lucide-react'
 import { PencilLine } from 'lucide-react';
 // TRPC Import
 import {api} from "~/trpc/server";
-import ProfileKegiatanComp from "~/app/profil-kegiatan/[profileKegiatanId]/profileKegiatanComp";
+import ProfileKegiatanComp from "~/app/_components/profil-kegiatan/profile-kegiatan-comp";
 import EditProfileLembaga from "~/app/lembaga/profil-lembaga/_components/edit-profile-lembaga";
 import {getServerAuthSession} from "~/server/auth";
